+++ packages/react-router/lib/server-runtime/server.ts
@@ -1,5 +1,5 @@
 import type { StaticHandler, StaticHandlerContext } from "../router/router";
-import type { ErrorResponse, unstable_InitialContext } from "../router/utils";
+import type { ErrorResponse } from "../router/utils";
 import { unstable_RouterContextProvider } from "../router/utils";
 import {
   isRouteErrorResponse,
@@ -48,13 +48,13 @@ import { getManifestPath } from "../dom/ssr/fog-of-war";
 export type RequestHandler = (
   request: Request,
   loadContext?: MiddlewareEnabled extends true
-    ? unstable_InitialContext
-    : AppLoadContext
+    ? unstable_RouterContextProvider
+    : AppLoadContext,
 ) => Promise<Response>;
 
 export type CreateRequestHandlerFunction = (
   build: ServerBuild | (() => ServerBuild | Promise<ServerBuild>),
-  mode?: string
+  mode?: string,
 ) => RequestHandler;
 
 function derive(build: ServerBuild, mode?: string) {
@@ -71,7 +71,7 @@ function derive(build: ServerBuild, mode?: string) {
       if (serverMode !== ServerMode.Test && !request.signal.aborted) {
         console.error(
           // @ts-expect-error This is "private" from users but intended for internal use
-          isRouteErrorResponse(error) && error.error ? error.error : error
+          isRouteErrorResponse(error) && error.error ? error.error : error,
         );
       }
     });
@@ -86,7 +86,7 @@ function derive(build: ServerBuild, mode?: string) {
 
 export const createRequestHandler: CreateRequestHandlerFunction = (
   build,
-  mode
+  mode,
 ) => {
   let _build: ServerBuild;
   let routes: ServerRoute[];
@@ -127,24 +127,19 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
     };
 
     if (_build.future.unstable_middleware) {
-      if (initialContext == null) {
-        loadContext = new unstable_RouterContextProvider();
-      } else {
-        try {
-          loadContext = new unstable_RouterContextProvider(
-            initialContext as unknown as unstable_InitialContext
-          );
-        } catch (e) {
-          let error = new Error(
-            "Unable to create initial `unstable_RouterContextProvider` instance. " +
-              "Please confirm you are returning an instance of " +
-              "`Map<unstable_routerContext, unknown>` from your `getLoadContext` function." +
-              `\n\nError: ${e instanceof Error ? e.toString() : e}`
-          );
-          handleError(error);
-          throw error;
-        }
+      if (
+        initialContext &&
+        !(initialContext instanceof unstable_RouterContextProvider)
+      ) {
+        let error = new Error(
+          "Invalid `context` value provided to `handleRequest`. When middleware " +
+            "is enabled you must return an instance of `unstable_RouterContextProvider` " +
+            "from your `getLoadContext` function.",
+        );
+        handleError(error);
+        throw error;
       }
+      loadContext = initialContext || new unstable_RouterContextProvider();
     } else {
       loadContext = initialContext || {};
     }
@@ -172,14 +167,17 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
     // When runtime SSR is disabled, make our dev server behave like the deployed
     // pre-rendered site would
     if (!_build.ssr) {
+      // Decode the URL path before checking against the prerender config
+      let decodedPath = decodeURI(normalizedPath);
+
       // When SSR is disabled this, file can only ever run during dev because we
       // delete the server build at the end of the build
       if (_build.prerender.length === 0) {
         // ssr:false and no prerender config indicates "SPA Mode"
         isSpaMode = true;
       } else if (
-        !_build.prerender.includes(normalizedPath) &&
-        !_build.prerender.includes(normalizedPath + "/")
+        !_build.prerender.includes(decodedPath) &&
+        !_build.prerender.includes(decodedPath + "/")
       ) {
         if (url.pathname.endsWith(".data")) {
           // 404 on non-pre-rendered `.data` requests
@@ -187,13 +185,13 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
             new ErrorResponseImpl(
               404,
               "Not Found",
-              `Refusing to SSR the path \`${normalizedPath}\` because \`ssr:false\` is set and the path is not included in the \`prerender\` config, so in production the path will be a 404.`
+              `Refusing to SSR the path \`${decodedPath}\` because \`ssr:false\` is set and the path is not included in the \`prerender\` config, so in production the path will be a 404.`,
             ),
             {
               context: loadContext,
               params,
               request,
-            }
+            },
           );
           return new Response("Not Found", {
             status: 404,
@@ -209,7 +207,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
     // Manifest request for fog of war
     let manifestUrl = getManifestPath(
       _build.routeDiscovery.manifestPath,
-      normalizedBasename
+      normalizedBasename,
     );
     if (url.pathname === manifestUrl) {
       try {
@@ -234,7 +232,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
       let singleFetchMatches = matchServerRoutes(
         routes,
         handlerUrl.pathname,
-        _build.basename
+        _build.basename,
       );
 
       response = await handleSingleFetchRequest(
@@ -244,7 +242,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
         request,
         handlerUrl,
         loadContext,
-        handleError
+        handleError,
       );
 
       if (_build.entry.module.handleDataRequest) {
@@ -259,7 +257,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
             getSingleFetchRedirect(
               response.status,
               response.headers,
-              _build.basename
+              _build.basename,
             );
 
           if (request.method === "GET") {
@@ -275,12 +273,12 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
               result,
               request.signal,
               _build.entry.module.streamTimeout,
-              serverMode
+              serverMode,
             ),
             {
               status: SINGLE_FETCH_REDIRECT_STATUS,
               headers,
-            }
+            },
           );
         }
       }
@@ -297,7 +295,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
         matches.slice(-1)[0].route.id,
         request,
         loadContext,
-        handleError
+        handleError,
       );
     } else {
       let { pathname } = url;
@@ -320,7 +318,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
         loadContext,
         handleError,
         isSpaMode,
-        criticalCss
+        criticalCss,
       );
     }
 
@@ -339,7 +337,7 @@ export const createRequestHandler: CreateRequestHandlerFunction = (
 async function handleManifestRequest(
   build: ServerBuild,
   routes: ServerRoute[],
-  url: URL
+  url: URL,
 ) {
   if (build.assets.version !== url.searchParams.get("version")) {
     return new Response(null, {
@@ -404,7 +402,7 @@ async function handleSingleFetchRequest(
   request: Request,
   handlerUrl: URL,
   loadContext: AppLoadContext | unstable_RouterContextProvider,
-  handleError: (err: unknown) => void
+  handleError: (err: unknown) => void,
 ): Promise<Response> {
   let response =
     request.method !== "GET"
@@ -415,7 +413,7 @@ async function handleSingleFetchRequest(
           request,
           handlerUrl,
           loadContext,
-          handleError
+          handleError,
         )
       : await singleFetchLoaders(
           build,
@@ -424,7 +422,7 @@ async function handleSingleFetchRequest(
           request,
           handlerUrl,
           loadContext,
-          handleError
+          handleError,
         );
 
   return response;
@@ -438,30 +436,38 @@ async function handleDocumentRequest(
   loadContext: AppLoadContext | unstable_RouterContextProvider,
   handleError: (err: unknown) => void,
   isSpaMode: boolean,
-  criticalCss?: CriticalCss
+  criticalCss?: CriticalCss,
 ) {
   try {
-    let response = await staticHandler.query(request, {
+    let result = await staticHandler.query(request, {
       requestContext: loadContext,
-      unstable_respond: build.future.unstable_middleware
-        ? (ctx) => renderHtml(ctx, isSpaMode)
+      unstable_generateMiddlewareResponse: build.future.unstable_middleware
+        ? async (query) => {
+            try {
+              let innerResult = await query(request);
+              if (!isResponse(innerResult)) {
+                innerResult = await renderHtml(innerResult, isSpaMode);
+              }
+              return innerResult;
+            } catch (error: unknown) {
+              handleError(error);
+              return new Response(null, { status: 500 });
+            }
+          }
         : undefined,
     });
-    // while middleware is still unstable, we don't run the middleware pipeline
-    // if no routes have middleware, so we still might need to convert context
-    // to a response here
-    return isResponse(response) ? response : renderHtml(response, isSpaMode);
+
+    if (!isResponse(result)) {
+      result = await renderHtml(result, isSpaMode);
+    }
+    return result;
   } catch (error: unknown) {
     handleError(error);
     return new Response(null, { status: 500 });
   }
 
   async function renderHtml(context: StaticHandlerContext, isSpaMode: boolean) {
-    if (isResponse(context)) {
-      return context;
-    }
-
-    let headers = getDocumentHeaders(build, context);
+    let headers = getDocumentHeaders(context, build);
 
     // Skip response body for unsupported status codes
     if (SERVER_NO_BODY_STATUS_CODES.has(context.statusCode)) {
@@ -507,7 +513,7 @@ async function handleDocumentRequest(
         state,
         request.signal,
         build.entry.module.streamTimeout,
-        serverMode
+        serverMode,
       ),
       renderMeta: {},
       future: build.future,
@@ -526,7 +532,7 @@ async function handleDocumentRequest(
         entryContext,
         loadContext as MiddlewareEnabled extends true
           ? unstable_RouterContextProvider
-          : AppLoadContext
+          : AppLoadContext,
       );
     } catch (error: unknown) {
       handleError(error);
@@ -541,7 +547,7 @@ async function handleDocumentRequest(
           errorForSecondRender = new ErrorResponseImpl(
             error.status,
             error.statusText,
-            data
+            data,
           );
         } catch (e) {
           // If we can't unwrap the response - just leave it as-is
@@ -552,7 +558,7 @@ async function handleDocumentRequest(
       context = getStaticContextFromError(
         staticHandler.dataRoutes,
         context,
-        errorForSecondRender
+        errorForSecondRender,
       );
 
       // Sanitize errors outside of development environments
@@ -577,7 +583,7 @@ async function handleDocumentRequest(
           state,
           request.signal,
           build.entry.module.streamTimeout,
-          serverMode
+          serverMode,
         ),
         renderMeta: {},
       };
@@ -590,7 +596,7 @@ async function handleDocumentRequest(
           entryContext,
           loadContext as MiddlewareEnabled extends true
             ? unstable_RouterContextProvider
-            : AppLoadContext
+            : AppLoadContext,
         );
       } catch (error: any) {
         handleError(error);
@@ -607,30 +613,47 @@ async function handleResourceRequest(
   routeId: string,
   request: Request,
   loadContext: AppLoadContext | unstable_RouterContextProvider,
-  handleError: (err: unknown) => void
+  handleError: (err: unknown) => void,
 ) {
   try {
     // Note we keep the routeId here to align with the Remix handling of
     // resource routes which doesn't take ?index into account and just takes
     // the leaf match
-    let response = await staticHandler.queryRoute(request, {
+    let result = await staticHandler.queryRoute(request, {
       routeId,
       requestContext: loadContext,
-      unstable_respond: build.future.unstable_middleware
-        ? (ctx) => ctx
+      unstable_generateMiddlewareResponse: build.future.unstable_middleware
+        ? async (queryRoute) => {
+            try {
+              let innerResult = await queryRoute(request);
+              return handleQueryRouteResult(innerResult);
+            } catch (error) {
+              return handleQueryRouteError(error);
+            }
+          }
         : undefined,
     });
 
-    if (isResponse(response)) {
-      return response;
-    }
-
-    if (typeof response === "string") {
-      return new Response(response);
-    }
-
-    return Response.json(response);
+    return handleQueryRouteResult(result);
   } catch (error: unknown) {
+    return handleQueryRouteError(error);
+  }
+
+  function handleQueryRouteResult(
+    result: Awaited<ReturnType<StaticHandler["queryRoute"]>>,
+  ) {
+    if (isResponse(result)) {
+      return result;
+    }
+
+    if (typeof result === "string") {
+      return new Response(result);
+    }
+
+    return Response.json(result);
+  }
+
+  function handleQueryRouteError(error: unknown) {
     if (isResponse(error)) {
       // Note: Not functionally required but ensures that our response headers
       // match identically to what Remix returns
@@ -639,9 +662,7 @@ async function handleResourceRequest(
     }
 
     if (isRouteErrorResponse(error)) {
-      if (error) {
-        handleError(error);
-      }
+      handleError(error);
       return errorResponseToJson(error, serverMode);
     }
 
@@ -650,7 +671,7 @@ async function handleResourceRequest(
       error.message === "Expected a response from queryRoute"
     ) {
       let newError = new Error(
-        "Expected a Response to be returned from resource route handler"
+        "Expected a Response to be returned from resource route handler",
       );
       handleError(newError);
       throw error;
@@ -663,13 +684,13 @@ async function handleResourceRequest(
 
 function errorResponseToJson(
   errorResponse: ErrorResponse,
-  serverMode: ServerMode
+  serverMode: ServerMode,
 ): Response {
   return Response.json(
     serializeError(
       // @ts-expect-error This is "private" from users but intended for internal use
       errorResponse.error || new Error("Unexpected Server Error"),
-      serverMode
+      serverMode,
     ),
     {
       status: errorResponse.status,
@@ -677,7 +698,7 @@ function errorResponseToJson(
       headers: {
         "X-Remix-Error": "yes",
       },
-    }
+    },
   );
 }