+++ packages/react-router-remix-routes-option-adapter/CHANGELOG.md
@@ -1,5 +1,30 @@
 # `@react-router/remix-config-routes-adapter`
 
+## 7.8.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@react-router/dev@7.8.0`
+
+## 7.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@react-router/dev@7.7.1`
+
+## 7.7.0
+
+### Minor Changes
+
+- Export `DefineRouteFunction` type alongside `DefineRoutesFunction` ([#13945](https://github.com/remix-run/react-router/pull/13945))
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@react-router/dev@7.7.0`
+
 ## 7.6.3
 
 ### Patch Changes